+++ src/Explorer/Controls/Settings/SettingsSubComponents/MaterializedViewComponent.tsx
@@ -1,4 +1,4 @@
-import { Link, PrimaryButton, Stack, Text } from "@fluentui/react";
+import { Link, Stack, Text } from "@fluentui/react";
 import React from "react";
 import * as ViewModels from "../../../../Contracts/ViewModels";
 import { MaterializedViewSourceComponent } from "./MaterializedViewSourceComponent";
@@ -8,22 +8,11 @@ export interface MaterializedViewComponentProps {
 }
 
 export const MaterializedViewComponent: React.FC<MaterializedViewComponentProps> = ({ collection }) => {
+  // If this container itself defines a materialized view, skip rendering this component.
   const isTargetContainer = !!collection?.materializedViewDefinition();
-  const materializedViews = collection?.materializedViews() ?? [];
-
   if (isTargetContainer) {
     return null;
   }
-
-  const showEditor = materializedViews.length > 0;
-  const jsonValue = JSON.stringify(
-    materializedViews.map((view) => ({
-      name: view.id,
-      definition: view.id,
-    })),
-    null,
-    2,
-  );
 
   return (
     <Stack tokens={{ childrenGap: 12 }} styles={{ root: { maxWidth: 600 } }}>
@@ -34,14 +23,7 @@ export const MaterializedViewComponent: React.FC<MaterializedViewComponentProps>
         </Link>
         <Text>about how to define materialized views and how to use them.</Text>
       </Stack>
-
-      {showEditor && <MaterializedViewSourceComponent jsonValue={jsonValue} />}
-
-      <PrimaryButton
-        text="Add view"
-        styles={{ root: { width: "fit-content" } }}
-        onClick={() => console.log("Add view clicked")}
-      />
+      <MaterializedViewSourceComponent collection={collection} />
     </Stack>
   );
 };
